requirements.txt
Django==1.6.5
argparse==1.2.1
django-allauth==0.16.1
django-bootstrap-form==3.1
django-crispy-forms==1.4.0
django-passwords==0.2.0
oauthlib==0.6.3
python-openid==2.2.5
requests==2.3.0
requests-oauthlib==0.4.1
wsgiref==0.1.2
